requirements.txt
ttkbootstrap>=1.10.1
pystray>=0.19.5
Pillow>=10.0.0
pyinstaller>=6.0.0
